migration: VacationInfoRepositoryTest
